backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==24.3
channels==4.3.1
channels-redis==4.3.0
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.8
boto3==1.35.28
Pillow==10.4.0
django-storages==1.14.4
python-decouple==3.8
